requirements.txt
# UI Framework
streamlit==1.31.0

# Google Gemini AI
google-generativeai==0.8.3

# MCP Protocol
mcp==1.0.0
pydantic==2.5.3

# Utilities
python-dotenv==1.0.0
httpx==0.27.0
uvicorn==0.27.0
fastapi==0.109.0
python-dateutil==2.8.2

# Google Calendar Integration
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.115.0
pytz==2024.1